requirements.txt
# AWS Strands Agent SDK
strands-agents>=0.1.0

# MCP SDK for Python (required for MCP client transports)
mcp>=1.0.0

# AWS Bedrock integration (default LLM provider for Strands)
boto3>=1.34.0

# Environment variable management
python-dotenv>=1.0.0

# Document generation (for FTR outputs)
python-docx>=1.1.0
openpyxl>=3.1.0

# HTTP client for API calls
httpx>=0.27.0
